Constraint: arrive after depart
Version 0.4.1 generates 1,895,469 valid flight records over a 90 day
period with 667 rejected with flights arriving before they depart.
Raised issue #2 to address this, for now we drop them from the output.
The errors don’t start immediately so it could be an error in the input
timetable.

This is 21,060 flights/day which is in the ballpark of Star Alliance
claiming they ran 18,043 flights/day in 2013.